Extract the IPython token (if there is one)
New versions of the Jupyter Notebook provide a token that a user must
provide to access the notebook. This allows one to have some basic level
of security when working with the notebook as others can't get access
without the token. However, this means we have to change our parsing of
the Jupyter log output. We do this in such a way that if an old
notebook without the token is provided it is a non-issue for us and can
still be handled. This token is then written into the set of environment
variables that a user gets when connecting with the login node.