Add LSF configuration in addition to SGE
This makes sure that ipyparallel is configured for use with LSF. The SGE
configuration is still included as well. Though as Janelia is moving to
LSF, we need to start configure it as well. Once SGE is dropped at
Janelia, we will probably want to drop it here as well.