Disable LSF job emails in all cases
It seems that LSF likes to send per job emails about the job status.
This has the potential to flood one's email account and generally make
it unusable. So simply set an environment variable to disable this
behavior in all cases. A user can easily override this behavior for a
job or session by unsetting this variable before submitting a job.