Use our own lsf-drmaa
As we need a copy of `lsf-drmaa`, which is compatible with our glibc
requirements, we built our own copy and added it to the `nanshe`
channel. Originally we built this for CentOS 6 (glibc 2.12)
compatibility. However as the cluster has CentOS 5 (glibc 2.3), we
decided to build `lsf-drmaa` on this older target instead for increased
flexibility. Hence we change our install scripts for the workflow to
install our `conda` package of `lsf-drmaa` and make sure that is the one
we use with the Python `drmaa` library. This way we can use `lsf-drmaa`
inside our CentOS 6 container even though the cluster is now running on
CentOS 7 (glibc 2.17).