Set JUPYTER_RUNTIME_DIR explicitly
Based on advice from our cluster administrators, we should not allow
Jupyter to write to the `XDG_RUNTIME_DIR`, which is what it will choose
to do with Jupyter's runtime directory. To fix this, we tried seeing
where Jupyter would pick to write to if `XDG_RUNTIME_DIR` were unset. In
this case, Jupyter picked `~/.local/share/jupyter/runtime`. So we
updated our scripts to simply set `JUPYTER_RUNTIME_DIR` to this same
path. Should be easily accessible both under normal circumstances and
within Singularity containers.