Use SciComp installed copy of our lsf-drmaa
As SciComp has nicely agreed to install our copy of lsf-drmaa in a
global directory we can access regardless of user, switch to using that
instead of installing our own copy. This copy was built in a CentOS 5
container so supports glibc 2.3+, which is compatible with the glibc 2.3
install of LSF (oldest glibc supported by LSF on our cluster). Also will
work great with our CentOS 6 based containers, which have glibc 2.12.
Not to mention this continues to work fine under ScientificLinux 7,
which uses glibc 2.17. Hence we drop our lsf-drmaa install and simply
rely on this external copy. Will make it easier to simplify our
requirements as we make more use of Singularity containers on the
cluster.