Ensure `~/.bash_profile` exists
As we no longer write a blank line into `~/.bash_profile` and we do need
to make sure it exists so that searching its contents work, make sure to
run `touch` on `~/.bash_profile`. This won't modify the contents of
`~/.bash_profile`. Though it will make sure it always exists.