Download and convert Docker image to Singularity
Drops all of the steps related to cloning the GitHub repo and building
the associated Conda package as we are able to reuse the prebuilt Docker
image that already does this on the cluster. So instead download and
convert the Docker image to a Singularity image in the user's home
directory if it has not already been done before.