Startup Singularity container for users
As the Singularity container can be started up easily and the notebook
accessed remotely without too much difficulty, clean out all of the
existing content in the startup script. Replace this with running the
Singularity container in the current directory. Make sure to bind the
`/misc` directory for access to DRMAA for launching cluster jobs and the
`/scratch` directory for access to node local scratch space for Dask.